Fix filled glyph issue
We're seeing an issue that's most easily demonstrated by the following
lowercase "o" shape:

......
..##..
.#..#.
.#..#.
..##..
......

Our previous approach contour-ized this as two contours: a clockwise
ring around the outside and a counterclockwise ring around the inside.
This works well in most cases, but not in the case of our lowercase "o"
there. Many (but not all) font renderers were seemingly ignoring the
inner contour and treating the figure as completely filled.

I believe (+ testing indicates) this issue only occurs when an inner
contour shares ALL of its points with another contour.

Originally I wanted to solve this by detecting this case and adding an
extra dummy point arbitrarily between two points, but there's a few
additional issues to consider:
- I was kind of dreading writing this detection code. I didn't end up
  writing it, but I'm having trouble thinking of how to do it better
  than O(mn²) where n is the number of contours and m is the number of
  pts per contour. Granted n's fairly small, but it's gross overhead on
  every glyph.
- Even after detecting this, there's no guarantee there will even be
  room for a dummy point to fit! Contour coordinates are integers: with
  a 1x1 empty area there's nowhere to fit another dummy point.
  As a separate issue, we're gonna have to scale up the glyphs anyway..
  maybe we try to unconditionally scale up to ensure there's room?
  There _is_ an upper bound to font size.. we _probably_ won't hit that?
  Maybe we catch this 1x1 case and fall back to a different wrapping
  technique?

...Say, what would that different wrapping technique look like, anyway?

If instead of treating the "o" above as two rings we treated it as
four disjoint polygons, this would solve our problem. How do we make
that happen? Turns out it's as easy as swapping our contour walking
behavior: instead of preferring to turn left at a catty-corner, we
prefer turning right:

........
.###....
.###....
.###....
...^###.
....###.
....###.
........

Previously we'd turn left at the arrow to capture the entire figure as
one contour, but now we turn right to only capture the bottom rect
(we pick up the top rect as a separate contour in a different pass).

This is what this commit changes: we're now unconditionally forming
more, smaller contours.

The tradeoff here is file size: I believe this strategy necessarily
results in an as-large or larger file size. Consider the "o" above:
we previously described one contour of 8 points and one contour of
4 points for a total of 12, but we're now describing four contours of
4 points each for a total of 16. In practical terms, the small font I've
been building this library for went from 6944 bytes to 7460 bytes.
500 bytes isn't the end of the world, +7% filesize isn't such an issue
(this commit message is six times that size...). But it's something to
be aware of. If filesize is a concern later, we can revisit the
fallback approach.

This bug wasn't a design oversight or typo or mistake really. It was
more of a hidden requirement. Given that, it's wild to me that this was
a one-line change to fix! (give or take tests.) I love it.